+++ openrlhf/models/actor.py
@@ -136,6 +136,7 @@ class Actor(nn.Module):
             "eos_token_id": kwargs.get("eos_token_id"),
             "pad_token_id": kwargs.get("pad_token_id"),
             "min_new_tokens": kwargs.get("min_new_tokens", 1),
+            "synced_gpus": kwargs.get("synced_gpus", False),
         }
 
         if kwargs.get("max_new_tokens", None):
